base/modules/utils/root/requirements.txt
bson==0.5.10
cbor2==5.1.2
kombu==5.0.0
msgpack==1.0.0
py3-bencode==0.0.3
py-ubjson==0.16.1
pyrabbit==1.1.0
pyyaml==5.3.1
sexpdata==0.0.3
toml==0.10.1
xmltodict==0.12.0